Major Updates

+++ main.py
@@ -1,58 +1,56 @@
+import pandas as pd
+import numpy as np
+import matplotlib.pyplot as plt
+from datetime import datetime as dt
+from pprint import pprint
 import requests
 import json
-import time
-import numpy as np
-import pandas as pd
+import os
 
 
 class CryptoBot:
-    def __init__(self, pair):
-        self.pair = pair
-        self.coin = 0
-        self.crypto_data = self.load_crypto_data()
-        self.trades = []
-        self.last_trade = None
-        self.balance = 0
+    def __init__(self, coin, time_interval, balance):
+        self.coin = coin
+        self.balance = balance
+        self.time = time_interval
+        self.last_transaction = None
 
     def save_crypto_data(self, data):
         '''
-        Saves the data in respective coin. 
+        Saves the data in respective coin.
         If the file doesn't exist, it creates and fills with default data.
         Updates new data of coin
         '''
-
         # If file not found, creates file and enters default value
-        try:
-            with open('data.json', 'r') as f:
-                all_data = json.load(f)
-        except:
+        if not os.path.exists('data.json'):
             with open('data.json', 'w') as f:
-                all_data = {"balance": 0, "crypto": {}}
-                json.dump(all_data, f, indent=4)
+                empty_data = {}
+                json.dump(empty_data, f, indent=4)
 
-        # Enters data to specific coin
-        all_data = {}
-        with open('data.json', 'r') as f:
+        with open('data.json', 'r+') as f:
             all_data = json.load(f)
-
-        with open('data.json', 'w') as f:
-            all_data["crypto"][self.pair] = data
+            if self.coin in all_data:
+                all_data[self.coin] = data
+            else:
+                all_data.update(data)
+            f.seek(0)
             json.dump(all_data, f, indent=4)
 
     def load_crypto_data(self):
         '''
-        Loads the data for a coin. 
-        If not found, it generates a new one for it. 
+        Loads the data for a coin.
+        If not found, it generates a new one for it.
         '''
-        data = {}
-        try:
+
+        if os.path.exists('data.json'):
             with open('data.json', 'r') as f:
                 data = json.load(f)
-                self.balance = data.get("balance")/len(data.get("crypto"))
-                data = self.crypto_data = data["crypto"][self.pair]
-                self.coin = data.get("coins")
-                self.trades = data.get("trades")
-        except:
+                if self.coin in data:
+                    data = data[self.coin]
+                else:
+                    data = self.make_crypto_data()
+                    self.save_crypto_data(data)
+        else:
             data = self.make_crypto_data()
             self.save_crypto_data(data)
         return data
@@ -61,97 +59,101 @@ class CryptoBot:
         '''
         Generates new data for new crypto coin
         '''
-        data = {
+        data = {}
+        data[self.coin] = {
             'trades': [],
+            'balance': self.balance,
             'coins': 0.0,
-            "worth": 0.0,
         }
         return data
 
-    def get_coin_data(self):
-        try:
-            data = requests.get(
-                f"https://public.coindcx.com/market_data/candles?pair={self.pair}&interval=1m").json()
-            self.last_trade = data[0]
-            data = data[1:101]
-            data = pd.DataFrame(data)
-            data = np.array(data)
-            return data
-        except:
-            return np.array([])
-
-    def get_trend(self):
-        coin_data = self.get_coin_data()
-        if len(coin_data):
-            high_max = (np.max(coin_data, axis=0))[1]
-            low_min = (np.min(coin_data, axis=0))[2]
-            print(high_max, low_min, self.last_trade.get("close"))
-            if self.last_trade.get("high") > high_max:
-                return "UPTREND"
-            elif self.last_trade.get("low") < low_min:
-                return "DOWNTREND"
-            else:
-                return "NOTREND"
-
-    def get_trade_data(self, trade, amount):
-        '''
-        Generates and saves Trade data to JSON file. 
-        '''
+    def make_trade_data(self, price, trade, amount):
         transaction = {
-            "time-stamp": str(int(time.time())),
-            "price_inr": self.last_trade.get("close"),
+            "time-stamp": str(dt.now()),
+            "price_inr": price,
             "trade": trade,
             "amount": amount
         }
-        print('TRADE:')
-        print(json.dumps(transaction, indent=4))
         return transaction
-        # self.crypto_data["trades"].append(transaction)
-        # if trade == "BUY":
-        #     self.crypto_data["coins"] = self.coin + \
-        #         (amount/self.last_trade.get("close"))
-        #     self.crypto_data["worth"] = self.crypto_data.get("worth") + amount
-        #     self.balance = self.balance - amount
-        # elif trade == "SELL":
-        #     self.crypto_data["coins"] = self.coin - \
-        #         (amount/self.last_trade.get("close"))
-        #     self.crypto_data["worth"] = self.crypto_data.get(
-        #         "worth") - (self.crypto_data["coins"]*self.last_trade.get("close"))
-        #     self.balance = self.balance + self.crypto_data["coins"] * \
-        #         self.last_trade.get("close")
-        # self.save_crypto_data(self.crypto_data)
 
-        # with open('data.json', 'r') as f:
-        #     all_data = json.load(f)
+    def get_coin_data(self):
+        data = requests.get(
+            f"https://public.coindcx.com/market_data/candles?pair={self.coin}&interval=1m&limit={self.time*2}").json()
+        data = pd.DataFrame(data, columns=["close"])
+        data = np.array(data)
+        data = data.flatten()
+        return np.flipud(data)
 
-        # with open('data.json', 'w') as f:
-        #     all_data["balance"] = self.balance
+    def get_coin_mean(self, data):
+        avg = np.array([])
+        for id in range(self.time):
+            arr = data[id:id+self.time]
+            avg = np.append(avg, np.mean(arr), axis=None)
+        data = data[time-1:-1]
+        return avg
 
-    def set_buy_data(self, tansaction):
-        pass
+    def get_mean_slope(self, data):
+        data = np.flipud(data)
+        slope = []
+        last = data[0]
+        for id in range(1, len(data), 10):
+            m = id/(last - data[id])
+            slope.append(m)
+        return np.flipud(np.array(slope))
 
-    def buy_crypto(self):
-        transaction = self.get_trade_data("BUY", self.balance)
-        self.set_buy_data(transaction)
+    def check_buy(self, coin_mean_slope):
+        check = 0
+        for id in range(3):
+            if coin_mean_slope[len(coin_mean_slope)-id - 1] > 0:
+                check += 1
+        if check == 3:
+            return "BUY"
+        return "IDLE"
 
-    def sell_crypto(self):
-        self.get_trade_data("SELL", 0.0)
+    def graph(self, price_data, mean_data):
+        plt.clf()
+        plt.plot(np.arange(start=0, stop=len(
+            price_data), step=1), price_data)
+        plt.plot(np.arange(start=0, stop=len(mean_data)*2, step=2),
+                 mean_data, c="red", linestyle="--")
+        plt.xlabel("Time")
+        plt.ylabel("Closing Price")
+        plt.title("Closing Price Data")
+        plt.pause(5)
+
+    def make_buy(self, price):
+        coin_data = self.load_crypto_data()
+
+        self.balance = coin_data.get("balance")
+        amount = self.balance/price
+        trade = self.make_trade_data(price, "BUY", amount)
+
+        prev_trades = coin_data.get("trades")
+        prev_trades.append(trade)
+        coin_data["trades"] = prev_trades
+        coin_data["coins"] = amount
+        coin_data["balance"] = coin_data.get("balance") - self.balance
+        print(coin_data)
 
     def driver(self):
         while True:
-            crypto_data = self.load_crypto_data()
-            trend = self.get_trend()
-            print(trend)
-            if trend == "UPTREND" and self.coin == 0.0:
-                # Buy
-                self.buy_crypto()
-            elif trend == "DOWNTREND" and self.coin > 0.0:
-                # Sell
-                self.sell_crypto()
-            time.sleep(20)
+            coin_data = self.get_coin_data()
+            coin_avg = self.get_coin_mean(coin_data)
+            coin_avg_slope = self.get_mean_slope(coin_avg)
+            current_state = self.check_buy(coin_avg_slope)
+            if current_state == "BUY":
+                self.make_buy(coin_data[-1])
+            print(current_state)
+            self.graph(coin_data, coin_avg)
 
 
 if __name__ == "__main__":
-    btc = CryptoBot("I-BTC_INR")
-    # print(btc.crypto_data)
-    btc.driver()
+    print("Starting...\n")
+    time = 120
+    # coin = "I-BTC_INR"
+    # coin = "I-MATIC_INR"
+    coin = "I-MANA_INR"
+    balance = 1000.0
+    bot = CryptoBot(coin, time, balance)
+    # print(bot.load_crypto_data())
+    bot.driver()
